requirements.txt
Django==5.0.3
python-nmap==0.7.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
scapy==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
